+++ docs/source/conf.py
@@ -22,7 +22,7 @@ copyright = '2022, Trey Hilyard'
 author = 'Trey Hilyard'
 
 # The full version, including alpha/beta/rc tags
-release = '1.0.0'
+release = '4.1.1'
 
 
 # -- General configuration ---------------------------------------------------